+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Fixed
+- Now sets the default language for the (upcoming) XMOD models.
+
+
 ## [v6.1.1] - 2023-01-02
 ### Fixed
 - Fixed error where some tokenizers did not have special token IDs registered.